RELATED: ONE-4163 Add locale option to DateFilter component
