+++ stories/__screenshots__/storybook_Internal_DateFilter_DateFilter_-_full-featured_opened_0_screenshot-wrapper_0_desktop.png
